requirements.txt
streamlit
tensorflow
opencv-python
pillow
streamlit-cropper
